+++ package/confd/confd.mk
@@ -10,7 +10,7 @@ CONFD_SITE_METHOD = local
 CONFD_SITE = $(BR2_EXTERNAL_INFIX_PATH)/src/confd
 CONFD_LICENSE = BSD-3
 CONFD_LICENSE_FILES = LICENSE
-CONFD_DEPENDENCIES = augeas jansson libite sysrepo
+CONFD_DEPENDENCIES = augeas jansson libite sysrepo libsrx
 CONFD_AUTORECONF = YES
 
 define CONFD_INSTALL_EXTRA
