+++ package/statd/statd.mk
@@ -3,11 +3,12 @@
 # statd
 #
 ################################################################################
+
 STATD_VERSION = 1.0
 STATD_LICENSE = MIT
 STATD_SITE_METHOD = local
 STATD_SITE = $(BR2_EXTERNAL_INFIX_PATH)/src/statd
-STATD_DEPENDENCIES = sysrepo libev jansson
+STATD_DEPENDENCIES = sysrepo libev libsrx jansson
 
 define STATD_BUILD_CMDS
 	$(TARGET_MAKE_ENV) $(TARGET_CONFIGURE_OPTS) $(MAKE) -C $(@D) \
