+++ src/confd/src/confd/Makefile.am
@@ -6,20 +6,22 @@ plugin_LTLIBRARIES      = confd-plugin.la
 
 confd_plugin_la_LDFLAGS = -module -avoid-version -shared
 
-confd_plugin_la_CFLAGS  = \
-	$(augeas_CFLAGS) \
-	$(glib_CFLAGS) \
-	$(jansson_CFLAGS) \
-	$(libite_CFLAGS) \
-	$(sysrepo_CFLAGS) \
+confd_plugin_la_CFLAGS  =			\
+	$(augeas_CFLAGS)			\
+	$(glib_CFLAGS)				\
+	$(jansson_CFLAGS)			\
+	$(libite_CFLAGS)			\
+	$(sysrepo_CFLAGS)			\
+	$(libsrx_CFLAGS)			\
 	$(CFLAGS)
 
-confd_plugin_la_LIBADD  = \
-	$(augeas_LIBS) \
-	$(glib_LIBS) \
-	$(jansson_LIBS) \
-	$(libite_LIBS) \
-	$(sysrepo_LIBS)
+confd_plugin_la_LIBADD  =			\
+	$(augeas_LIBS)				\
+	$(glib_LIBS)				\
+	$(jansson_LIBS)				\
+	$(libite_LIBS)				\
+	$(sysrepo_LIBS)				\
+	$(libsrx_LIBS)
 
 rauc_installer_sources = \
 	rauc-installer.c \
@@ -29,29 +31,21 @@ BUILT_SOURCES = $(rauc_installer_sources)
 EXTRA_DIST = ../lib/de.pengutronix.rauc.Installer.xml
 
 $(rauc_installer_sources): ../lib/de.pengutronix.rauc.Installer.xml
-	$(AM_V_GEN) gdbus-codegen \
-		--generate-c-code rauc-installer \
-		--c-generate-autocleanup all \
-		--c-namespace Rauc \
-		--interface-prefix de.pengutronix.rauc. \
+	$(AM_V_GEN) gdbus-codegen			\
+		--generate-c-code rauc-installer	\
+		--c-generate-autocleanup all		\
+		--c-namespace Rauc			\
+		--interface-prefix de.pengutronix.rauc.	\
 		$<
 
 nodist_confd_plugin_la_SOURCES = \
 	$(rauc_installer_sources)
 
-confd_plugin_la_SOURCES = \
-	ietf-interfaces.c \
-	ietf-system.c \
-	infix-dhcp.c \
-	infix-factory.c \
-	infix-system-software.c \
-	\
-	../lib/common.h \
-	../lib/helpers.c	../lib/helpers.h \
-	../lib/lyx.c		../lib/lyx.h \
-	../lib/srx_module.c	../lib/srx_module.h \
-	\
-	core.c		core.h \
-	dagger.c	dagger.h \
-	srx_val.c	srx_val.h \
-	systemv.c	systemv.h
+confd_plugin_la_SOURCES =			\
+	ietf-interfaces.c			\
+	ietf-system.c				\
+	infix-dhcp.c				\
+	infix-factory.c				\
+	infix-system-software.c			\
+	core.c		core.h			\
+	dagger.c	dagger.h
