+++ src/libsrx/LICENSE
@@ -1,4 +1,3 @@
-Copyright (c) 2020 Denis Kalashnikov
 Copyright (c) 2023 The KernelKit Authors
 All rights reserved.
