+++ src/libsrx/Makefile.am
@@ -1,0 +1,3 @@
+SUBDIRS             = src
+dist_doc_DATA       = README.md LICENSE
+ACLOCAL_AMFLAGS     = -I m4
