+++ src/libsrx/README.md
@@ -1,0 +1,6 @@
+sysepo extensions
+=================
+
+This library contains extensions and helper APIs for sysrepo plugins.
+It is currently used by confd and statd in Infix.
+
