+++ src/libsrx/autogen.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+autoreconf -W portability -visfm
