+++ src/libsrx/configure.ac
@@ -1,0 +1,16 @@
+AC_INIT(libsrx, 1.0.0, https://github.com/kernelkit/infix/issues)
+AC_CONFIG_AUX_DIR(aux)
+AM_INIT_AUTOMAKE([1.11 foreign])
+AM_SILENT_RULES([yes])
+
+AC_CONFIG_SRCDIR(src/lyx.h)
+AC_CONFIG_HEADERS(config.h)
+AC_CONFIG_FILES([Makefile src/Makefile src/libsrx.pc])
+AC_CONFIG_MACRO_DIR([m4])
+
+AC_PROG_CC
+AC_PROG_INSTALL
+AM_PROG_AR
+LT_INIT
+
+AC_OUTPUT
