+++ src/libsrx/src/helpers.c
