+++ src/libsrx/src/helpers.h
