+++ src/libsrx/src/libsrx.pc.in
@@ -1,0 +1,12 @@
+prefix=@prefix@
+exec_prefix=@exec_prefix@
+libdir=@libdir@
+includedir=@includedir@
+
+Name: @PACKAGE@
+Description: Extensions and helper APIs for sysrepo plugins
+Version: @VERSION@
+Requires:
+Libs: -L${libdir} -lsrx
+Cflags: -I${includedir}
+
