+++ src/libsrx/src/lyx.c
