+++ src/libsrx/src/lyx.h
