+++ src/libsrx/src/srx_val.c
@@ -1,8 +1,9 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
-
+#include <errno.h>
 #include <stdarg.h>
-#include "core.h"
-#include "../lib/common.h"
+#include <sysrepo/values.h>
+
+#include "common.h"
 
 sr_error_t srx_get_diff(sr_session_ctx_t *session, struct lyd_node **treep)
 {
